+++ src/app/active-mode/ActiveMode.m.scss
@@ -7,6 +7,7 @@
 .activityColumn {
   margin-top: 0;
   top: $header-height;
+  height: 100vh;
 
   :global(.store-cell) {
     padding: 0;
@@ -50,11 +51,11 @@
 
 .activityInventory {
   padding-top: 4px;
-  
+
   @include phone-portrait {
     padding-top: 0;
   }
-  
+
   :global(.store-row) {
     grid-template-columns:
       repeat(1, calc(6px + var(--character-column-width) + var(--column-padding)))
