+++ src/app/active-mode/Views/FarmingView.m.scss
@@ -1,9 +1,26 @@
 @import '../../variables.scss';
 
-.newItems {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(var(--item-size), 1fr));
-  & > div {
-    padding-bottom: 7px;
+.noSearches {
+  grid-column: 1 / -1;
+  margin: 10px 0;
+}
+
+.message {
+  text-align: center;
+  padding: 10px;
+  border-radius: 5px;
+  background-color: rgba(50, 0, 0, 0.8);
+  opacity: 0.6;
+}
+
+.options {
+  display: flex;
+}
+
+.applySearch {
+  color: #999;
+  cursor: pointer;
+  :hover {
+    color: $orange;
   }
 }
