+++ src/app/active-mode/Views/FarmingView.m.scss.d.ts
@@ -1,7 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'newItems': string;
+  'applySearch': string;
+  'message': string;
+  'noSearches': string;
+  'options': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
