+++ src/app/active-mode/Views/FarmingView.tsx
@@ -1,62 +1,83 @@
-import styles from 'app/active-mode/Views/FarmingView.m.scss';
+import { recentSearchesSelector } from 'app/dim-api/selectors';
 import CollapsibleTitle from 'app/dim-ui/CollapsibleTitle';
+import Dropdown, { Option } from 'app/dim-ui/Dropdown';
+import { itemPop } from 'app/dim-ui/scroll';
 import { startFarming, stopFarming } from 'app/farming/actions';
 import { t } from 'app/i18next-t';
-import InventoryItem from 'app/inventory/InventoryItem';
-import { DimItem } from 'app/inventory/item-types';
-import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
+import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import { allItemsSelector } from 'app/inventory/selectors';
 import { DimStore } from 'app/inventory/store-types';
-import { RootState, ThunkDispatchProp } from 'app/store/types';
-import React, { useEffect, useState } from 'react';
-import { connect } from 'react-redux';
+import ItemActionsDropdown from 'app/item-actions/ItemActionsDropdown';
+import { searchFiltersConfigSelector } from 'app/search/search-filter';
+import { setSearchQuery } from 'app/shell/actions';
+import { AppIcon, searchIcon } from 'app/shell/icons';
+import _ from 'lodash';
+import React, { useEffect, useMemo, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import styles from './FarmingView.m.scss';
 
-interface ProvidedProps {
+type Props = {
   store: DimStore;
-}
+};
 
-interface StoreProps {
-  newItems: DimItem[];
-}
-
-const newItemsToShow = 8;
-
-function mapStateToProps(state: RootState): StoreProps {
-  const newItems = allItemsSelector(state).sort((a, b) => (a.id > b.id ? -1 : 1));
-  newItems.length = newItemsToShow;
-
-  return {
-    newItems,
-  };
-}
-
-type Props = ProvidedProps & StoreProps & ThunkDispatchProp;
-
-function FarmingView({ store, newItems, dispatch }: Props) {
+export default function FarmingView({ store }: Props) {
   const [isFarming, setIsFarming] = useState(false);
 
+  const allItems = useSelector(allItemsSelector);
+  const filters = useSelector(searchFiltersConfigSelector);
+  const savedSearches = useSelector(recentSearchesSelector).filter(({ saved }) => saved);
+  const [query, setQuery] = useState<string>(savedSearches?.[0].query);
+  const options: Option[] = savedSearches.map(({ query }) => ({
+    key: query,
+    content: query,
+    onSelected: () => setQuery(query),
+  }));
+
+  const filter = useMemo(() => query && filters(query), [filters, query]);
+  const items = useMemo(
+    () => _.sortBy(filter ? allItems.filter(filter) : allItems, ({ id }) => -id),
+    [allItems, filter]
+  );
+
+  const dispatch = useDispatch();
   useEffect(() => {
     dispatch((isFarming ? startFarming : stopFarming)(store.id));
   }, [dispatch, store, isFarming]);
 
+  items.length = Math.min(items.length, 8);
+
   return (
     <CollapsibleTitle
       title={t('ActiveMode.Farming')}
-      sectionId={'active-farming'}
+      sectionId={'active-filter'}
       defaultCollapsed={true}
     >
       <div className="dim-button bucket-button" onClick={() => setIsFarming(!isFarming)}>
         {!isFarming ? t('ActiveMode.FarmingStart') : t('ActiveMode.FarmingStart')}
       </div>
-      <div className={styles.newItems}>
-        {newItems.map((item) => (
-          <ItemPopupTrigger key={item.id} item={item}>
-            {(ref, onClick) => <InventoryItem item={item} innerRef={ref} onClick={onClick} />}
-          </ItemPopupTrigger>
-        ))}
+      <div className={styles.options}>
+        <Dropdown options={options}>{t('ActiveMode.ChangeFilter')}</Dropdown>
+        <ItemActionsDropdown filteredItems={items} searchActive={Boolean(query.length)} />
+        <div className={styles.applySearch} onClick={() => dispatch(setSearchQuery(query))}>
+          <AppIcon icon={searchIcon} />
+        </div>
+      </div>
+      <div className="sub-bucket">
+        {options.length ? (
+          items.map((item) => (
+            <ConnectedInventoryItem
+              key={item.index}
+              id={'farm-' + item.index}
+              item={item}
+              onClick={() => itemPop(item)}
+            />
+          ))
+        ) : (
+          <div className={styles.noSearches}>
+            <div className={styles.message}>{t('ActiveMode.NoSearches')}</div>
+          </div>
+        )}
       </div>
     </CollapsibleTitle>
   );
 }
-
-export default connect<StoreProps>(mapStateToProps)(FarmingView);
