+++ src/app/inventory/ConnectedInventoryItem.tsx
@@ -13,6 +13,7 @@ import { itemHashTagsSelector, itemInfosSelector } from './selectors';
 // Props provided from parents
 interface ProvidedProps {
   item: DimItem;
+  id?: string; // defaults to item.index - id is typically used for `itemPop`
   allowFilter?: boolean;
   ignoreSelectedPerks?: boolean;
   innerRef?: React.Ref<HTMLDivElement>;
@@ -56,6 +57,7 @@ type Props = ProvidedProps & StoreProps;
  */
 function ConnectedInventoryItem({
   item,
+  id,
   isNew,
   tag,
   notes,
@@ -71,6 +73,7 @@ function ConnectedInventoryItem({
   return (
     <InventoryItem
       item={item}
+      id={id}
       isNew={isNew}
       tag={tag}
       notes={notes}
