+++ src/app/inventory/InventoryItem.tsx
@@ -15,6 +15,8 @@ import TagIcon from './TagIcon';
 
 interface Props {
   item: DimItem;
+  /** Optional id, otherwise will use item.index */
+  id?: string;
   /** Show this item as new? */
   isNew?: boolean;
   /** User defined tag */
@@ -36,6 +38,7 @@ interface Props {
 
 export default function InventoryItem({
   item,
+  id,
   isNew,
   tag,
   notes,
@@ -112,7 +115,7 @@ export default function InventoryItem({
 
   return (
     <div
-      id={item.index}
+      id={id || item.index}
       onClick={enhancedOnClick}
       onDoubleClick={onDoubleClick}
       title={`${item.name}\n${item.typeName}`}
