+++ src/app/item-actions/ItemActionsDropdown.m.scss
@@ -1,0 +1,7 @@
+.dropdownButton {
+  margin: 0 4px 0 8px;
+}
+
+.storeIcon {
+  margin-right: 4px;
+}
