+++ src/app/item-actions/ItemActionsDropdown.m.scss.d.ts
@@ -1,7 +1,8 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'count': string;
+  'dropdownButton': string;
+  'storeIcon': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
