+++ src/app/loadout/auto-loadouts.ts
@@ -192,15 +192,9 @@ export function gatherEngramsLoadout(
 }
 
 /**
- * Move items matching the current search.
+ * Move a list of items to a store
  */
-export function searchLoadout(
-  allItems: DimItem[],
-  store: DimStore,
-  searchFilter: ItemFilter
-): Loadout {
-  let items = allItems.filter((i) => !i.location.inPostmaster && !i.notransfer && searchFilter(i));
-
+export function itemMoveLoadout(items: DimItem[], store: DimStore): Loadout {
   items = addUpStackables(items);
 
   const itemsByType = _.mapValues(
@@ -214,6 +208,21 @@ export function searchLoadout(
     .map((i) => convertToLoadoutItem(i, false));
 
   return newLoadout(t('Loadouts.FilteredItems'), finalItems);
+}
+
+/**
+ * Move items matching the current search.
+ */
+export function searchLoadout(
+  allItems: DimItem[],
+  store: DimStore,
+  searchFilter: ItemFilter
+): Loadout {
+  const items = allItems.filter(
+    (i) => !i.location.inPostmaster && !i.notransfer && searchFilter(i)
+  );
+
+  return itemMoveLoadout(items, store);
 }
 
 function limitToBucketSize(items: DimItem[], isVault: boolean) {
