+++ src/app/search/MainSearchBarActions.tsx
@@ -1,38 +1,18 @@
-import Dropdown, { Option } from 'app/dim-ui/Dropdown';
 import { t } from 'app/i18next-t';
-import { setItemNote } from 'app/inventory/actions';
-import { bulkLockItems, bulkTagItems } from 'app/inventory/bulk-actions';
 import { InventoryBuckets } from 'app/inventory/inventory-buckets';
 import { allItemsSelector, bucketsSelector, sortedStoresSelector } from 'app/inventory/selectors';
 import { DimStore } from 'app/inventory/store-types';
-import { searchLoadout } from 'app/loadout/auto-loadouts';
-import { applyLoadout } from 'app/loadout/loadout-apply';
-import { isPhonePortraitSelector, querySelector } from 'app/shell/selectors';
+import ItemActionsDropdown from 'app/item-actions/ItemActionsDropdown';
+import { querySelector } from 'app/shell/selectors';
 import { RootState, ThunkDispatchProp } from 'app/store/types';
 import { emptyArray, emptySet } from 'app/utils/empty';
 import React, { useMemo } from 'react';
 import { connect } from 'react-redux';
 import { useLocation } from 'react-router';
-import { CompareService } from '../compare/compare.service';
-import { isTagValue, itemTagSelectorList, TagValue } from '../inventory/dim-item-info';
 import { DimItem } from '../inventory/item-types';
-import {
-  AppIcon,
-  clearIcon,
-  compareIcon,
-  lockIcon,
-  stickyNoteIcon,
-  unlockedIcon,
-} from '../shell/icons';
-import { loadingTracker } from '../shell/loading-tracker';
 import { ItemFilter } from './filter-types';
 import styles from './MainSearchBarActions.m.scss';
 import { searchFilterSelector } from './search-filter';
-import './search-filter.scss';
-
-interface ProvidedProps {
-  onClear?(): void;
-}
 
 interface StoreProps {
   searchQuery: string;
@@ -40,10 +20,9 @@ interface StoreProps {
   stores: DimStore[];
   buckets?: InventoryBuckets;
   searchFilter: ItemFilter;
-  isPhonePortrait: boolean;
 }
 
-type Props = ProvidedProps & StoreProps & ThunkDispatchProp;
+type Props = StoreProps & ThunkDispatchProp;
 
 function mapStateToProps(state: RootState): StoreProps {
   return {
@@ -52,22 +31,13 @@ function mapStateToProps(state: RootState): StoreProps {
     searchFilter: searchFilterSelector(state),
     allItems: allItemsSelector(state),
     buckets: bucketsSelector(state),
-    isPhonePortrait: isPhonePortraitSelector(state),
   };
 }
 
 /**
  * The extra buttons that appear in the main search bar when there are matched items.
  */
-function MainSearchBarActions({
-  allItems,
-  stores,
-  buckets,
-  searchFilter,
-  searchQuery,
-  isPhonePortrait,
-  dispatch,
-}: Props) {
+function MainSearchBarActions({ allItems, buckets, searchFilter, searchQuery }: Props) {
   const location = useLocation();
   const onInventory = location.pathname.endsWith('inventory');
   const onProgress = location.pathname.endsWith('progress');
@@ -75,8 +45,8 @@ function MainSearchBarActions({
   const onVendors = location.pathname.endsWith('vendors');
   // We don't have access to the selected store so we'd match multiple characters' worth.
   // Just suppress the count for now
-  const showSearchCount = searchQuery && !onProgress && !onRecords && !onVendors;
   const showSearchActions = onInventory;
+  const showSearchCount = Boolean(searchQuery && !onProgress && !onRecords && !onVendors);
 
   const displayableBuckets = useMemo(
     () =>
@@ -100,128 +70,6 @@ function MainSearchBarActions({
     [displayableBuckets, showSearchCount, searchFilter, allItems]
   );
 
-  let isComparable = false;
-  if (filteredItems.length && !CompareService.dialogOpen) {
-    const type = filteredItems[0].typeName;
-    isComparable = filteredItems.every((i) => i.typeName === type);
-  }
-
-  const bulkTag = loadingTracker.trackPromise(async (selectedTag: TagValue) => {
-    // Bulk tagging
-    const tagItems = filteredItems.filter((i) => i.taggable);
-
-    if (isTagValue(selectedTag)) {
-      dispatch(bulkTagItems(tagItems, selectedTag));
-    }
-  });
-
-  const bulkLock = loadingTracker.trackPromise(async (selectedTag: TagValue) => {
-    // Bulk locking/unlocking
-    const state = selectedTag === 'lock';
-    const lockables = filteredItems.filter((i) => i.lockable);
-    dispatch(bulkLockItems(lockables, state));
-  });
-
-  const bulkNote = () => {
-    const note = prompt(t('Organizer.NotePrompt'));
-    if (note !== null) {
-      if (filteredItems.length) {
-        for (const item of filteredItems) {
-          dispatch(setItemNote({ itemId: item.id, note: note || undefined }));
-        }
-      }
-    }
-  };
-
-  const compareMatching = () => {
-    CompareService.addItemsToCompare(filteredItems, false);
-  };
-
-  // Move items matching the current search. Max 9 per type.
-  const applySearchLoadout = async (store: DimStore) => {
-    const loadout = searchLoadout(allItems, store, searchFilter);
-    dispatch(applyLoadout(store, loadout, true));
-  };
-
-  const bulkItemTags = itemTagSelectorList
-    .filter((t) => t.type)
-    .map((tag) => ({
-      ...tag,
-      label: t('Header.TagAs', { tag: t(tag.label) }),
-    }));
-  bulkItemTags.push({ type: 'clear', label: t('Tags.ClearTag'), icon: clearIcon });
-
-  const dropdownOptions: Option[] = showSearchActions
-    ? [
-        ...stores.map((store) => ({
-          key: `move-${store.id}`,
-          onSelected: () => applySearchLoadout(store),
-          disabled: !showSearchCount,
-          content: (
-            <>
-              <img src={store.icon} width="16" height="16" alt="" className={styles.storeIcon} />{' '}
-              {store.isVault
-                ? t('MovePopup.SendToVault')
-                : t('MovePopup.StoreWithName', { character: store.name })}
-            </>
-          ),
-        })),
-        { key: 'characters' },
-        {
-          key: 'compare',
-          onSelected: compareMatching,
-          disabled: !isComparable || !showSearchCount,
-          content: (
-            <>
-              <AppIcon icon={compareIcon} /> {t('Header.CompareMatching')}
-            </>
-          ),
-        },
-        {
-          key: 'note',
-          onSelected: () => bulkNote(),
-          disabled: !showSearchCount,
-          content: (
-            <>
-              <AppIcon icon={stickyNoteIcon} /> {t('Organizer.Note')}
-            </>
-          ),
-        },
-        {
-          key: 'lock',
-          onSelected: () => bulkLock('lock'),
-          disabled: !showSearchCount,
-          content: (
-            <>
-              <AppIcon icon={lockIcon} /> {t('Tags.LockAll')}
-            </>
-          ),
-        },
-
-        {
-          key: 'unlock',
-          onSelected: () => bulkLock('unlock'),
-          disabled: !showSearchCount,
-          content: (
-            <>
-              <AppIcon icon={unlockedIcon} /> {t('Tags.UnlockAll')}
-            </>
-          ),
-        },
-        { key: 'tags' },
-        ...bulkItemTags.map((tag) => ({
-          key: tag.type || 'default',
-          onSelected: () => tag.type && bulkTag(tag.type),
-          disabled: !showSearchCount,
-          content: (
-            <>
-              {tag.icon && <AppIcon icon={tag.icon} />} {tag.label}
-            </>
-          ),
-        })),
-      ]
-    : emptyArray();
-
   return (
     <>
       {showSearchCount && (
@@ -231,12 +79,7 @@ function MainSearchBarActions({
       )}
 
       {showSearchActions && (
-        <Dropdown
-          options={dropdownOptions}
-          kebab={true}
-          className={styles.dropdownButton}
-          offset={isPhonePortrait ? 10 : 3}
-        />
+        <ItemActionsDropdown filteredItems={filteredItems} searchActive={showSearchCount} />
       )}
     </>
   );
